# openapi-herramientas-equipos.yaml
openapi: 3.0.3
info:
  title: API de Herramientas y Equipos Retirados
  description: >
    Módulo de backend de la app Wifix (Fase 1). Cubre dos dominios del
    servicio de operaciones de campo:

      1. HERRAMIENTAS — almacenamiento de resultados de mediciones:
         medición de distancia, test de velocidad, mapa de calor WiFi,
         ping y traceroute.
      2. EQUIPOS RETIRADOS — registro de equipos retirados o reemplazados.

    CLAVE DE IDENTIFICACIÓN (FASE 1)
    ----------------------------------------------------------------------
    En la Fase 1 la clave práctica de cada registro es `accountNumber`:
    el número de cuenta del cliente que el técnico ingresa en la app.
    Los demás identificadores (`clientId`, `visitId`, `technicianId`) son
    opcionales: provendrán de un sistema upstream aún por definir y en la
    Fase 1 se completan con valores de prueba. `contractId` también es
    opcional y en una fase posterior será reemplazado por `taskId`.

    AUTENTICACIÓN
    ----------------------------------------------------------------------
    La app tiene login propio. POST /auth/login devuelve un JWT que el
    resto de endpoints exige en el header Authorization: Bearer.

    DATOS DE SISTEMAS EXTERNOS (MOCKUP)
    ----------------------------------------------------------------------
    Los campos que provienen de sistemas de la operadora (Comarch/TYTAN,
    FSM, ISP Monitor, ACS, TEC, RMS) se sirven en esta etapa con datos
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /accounts/{accountNumber}/client-profile:
    parameters:
      - $ref: '#/components/parameters/AccountNumberPath'
    get:
      tags: [Datos del Cliente]
      summary: Perfil del cliente
      description: >
        Nombres, dirección, teléfonos, plan y velocidad contratada
        (campos 1, 2, 3, 5). **Respuesta compuesta** (ADR-07): la identidad sale
        de FSM `POST /account/process` (1 llamada upstream) y el plan/velocidad
        siguen viniendo del mock de Comarch, porque `fsm-data-ms` no los expone.
        El bloque `sources` dice, campo por campo, de dónde salió cada dato.
        Lo editado con el PUT manda sobre FSM y se marca como `MOCK`.
      operationId: getClientProfile
      parameters:
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: >-
            La cuenta no existe en FSM (`/account/process` devolvió una lista
            de órdenes vacía).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
    put:
      tags: [Datos del Cliente]
      summary: Actualizar datos del cliente
      description: >
        Módulo de actualización de datos (campo 4). **No cambia con la
        integración FSM**: `fsm-data-ms` es de solo lectura, así que la edición
        se guarda en Comarch y el cambio NO viaja a la operadora. Devuelve el
        mismo shape que el GET, con `sources` en `MOCK` para lo editado.
      operationId: updateClientProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientProfileUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /accounts/{accountNumber}/contract-status:
    get:
      tags: [Datos del Cliente]
      summary: Status del cliente por contrato/cuenta
      description: >
        Nombre del cliente y status de la cuenta (campo 7). Origen: FSM
        `POST /account/status` (código A/S/T/O/P) + `POST /account/process`
        (nombre y última orden). **2 llamadas upstream**, ambas cacheadas
        (status 5 min, process 60 s).


        FSM no devuelve "otras cuentas del mismo contrato": `accounts` trae
        exactamente un elemento hasta que exista ese endpoint.
      operationId: getContractStatus
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'

  # ====================================================================
  # DIAGNÓSTICO DE RED (campos 6, 8-14, 19-21)
  # ====================================================================
...
  /accounts/{accountNumber}/orders:
    get:
      tags: [Tareas y Visitas]
      summary: Órdenes de trabajo de la cuenta
      description: >
        Órdenes crudas normalizadas de FSM `POST /account/process`
        (base de los campos 1-3, 15 y 16; también sirve de diagnóstico).
        **1 llamada upstream.** No devuelve 404: una cuenta sin órdenes es un
        `orders: []` legítimo.
      operationId: getAccountOrders
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
        - name: estado
          in: query
          required: false
          description: >
            `Todas`: finalizadas y no finalizadas. `Pendientes`: solo las que
            aún no finalizan (`finish_date is null`).
          schema:
            type: string
            enum: [Todas, Pendientes]
            default: Todas
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountOrders'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por POST /auth/login.

  # --------------------------------------------------------------------
  # PARÁMETROS REUTILIZABLES
  # --------------------------------------------------------------------
  parameters:
...
    AccountNumberPath:
      name: accountNumber
      in: path
      required: true
      description: Número de cuenta del cliente.
      schema:
        type: string
...
    BrandHeader:
      name: X-Wifix-Brand
      in: header
      required: false
      description: >
        Marca comercial (realm de FSM) con la que se consulta la operadora.
        Si falta, el backend usa `FSM_DEFAULT_BRAND`. Internamente nunca se
        habla de "realm": los nombres `realm-ecommerce-callcenter-*` quedan
        encerrados en el backend y el técnico no los ve (ADR-01).
      schema:
        type: string
        enum: [telenews, seteinfo]
        example: telenews
    BrandQuery:
      name: brand
      in: query
      required: false
      description: >
        Override de `X-Wifix-Brand` por query. Pensado para el script de sondeo
        y soporte; la webapp usa el header.
      schema:
        type: string
        enum: [telenews, seteinfo]
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Token ausente, inválido o expirado / credenciales incorrectas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConnectorError:
      description: >
        Un sistema externo falló (5xx, timeout o JSON inválido). No es un
        problema de credenciales.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamAuthUnavailable:
      description: >
        El acceso a FSM no está disponible: token ausente, vencido o rechazado.


        **Es 503 a propósito, nunca 401.** La webapp cierra la sesión del
        técnico ante cualquier 401 (`api.js`), y un token caducado de un tercero
        no puede sacarlo al login en medio de una visita ni tumbarle Herramientas
        y Equipos Retirados, que no dependen de FSM.


        El cliente debe pintar un banner no bloqueante en el panel afectado y
        **no** cerrar sesión, no limpiar el token de Wifix y no recargar la app.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpstreamAuthError'

  # --------------------------------------------------------------------
  # ESQUEMAS
  # --------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Código de error legible por máquina.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Descripción del error legible por humanos.
          example: El campo accountNumber es obligatorio.
        details:
          type: array
          description: Detalle por campo, cuando aplica.
          items:
            type: object
            properties:
              field:
                type: string
              issue:
                type: string
        meta:
          type: object
          additionalProperties: true
          description: >
            Contexto adicional legible por máquina (integración, marca, motivo).
            Opcional y aditivo: el cliente nunca debe asumir que existe.
    UpstreamAuthError:
      type: object
      description: >
        Error de autenticación contra una API de terceros (hoy FSM). Siempre
        HTTP 503.
      required: [code, message]
      properties:
        code:
          type: string
          enum: [UPSTREAM_AUTH_ERROR]
        message:
          type: string
          description: Texto en español, listo para mostrar al técnico.
          example: >-
            El token de FSM (marca telenews) venció el 05/09/2026 21:59. Pide la
            renovación al contacto de la operadora.
        meta:
          type: object
          properties:
            integration:
              type: string
              example: FSM
            brand:
              type: string
              enum: [telenews, seteinfo]
            reason:
              type: string
              enum: [MISSING, EXPIRED, REJECTED]
              description: >
                `MISSING`: no hay acceso configurado en el servidor.
                `EXPIRED`: el token venció (detectado SIN salir a la red).
                `REJECTED`: la operadora lo rechazó (revocado o de otro realm).
            tokenExpiresAt:
              type: string
              format: date-time
              nullable: true
            retryable:
              type: boolean
              example: false
...
    ClientProfile:
      type: object
      properties:
        accountNumber:
          type: string
        fullName:
          type: string
          description: Nombres y apellidos (campo 1).
        address:
          type: string
          description: Dirección (campo 2).
        phones:
          type: array
          description: Teléfonos (campo 3).
          items:
            type: string
        planName:
          type: string
        contractedDownloadMbps:
          type: number
          format: double
          description: Velocidad de descarga contratada (campo 5).
        contractedUploadMbps:
          type: number
          format: double
          description: Velocidad de subida contratada (campo 5).
        email:
          type: string
          nullable: true
          description: Correo del cliente según FSM. Puede ser null.
        latitude:
          type: number
          format: double
          nullable: true
          description: >
            Coordenada del domicilio según la orden más reciente. Es la entrada
            del campo 6: evita que el técnico tenga que capturar GPS cuando la
            orden ya trae la coordenada.
        longitude:
          type: number
          format: double
          nullable: true
        sources:
          type: object
          description: >
            De dónde salió cada campo. `FSM` = dato real de la operadora,
            `MOCK` = simulado (o editado a mano con el PUT), `COMARCH` = TYTAN.
            La UI marca visualmente lo que no es `FSM`. Aditivo: quien lo ignore
            sigue funcionando.
          additionalProperties:
            type: string
            enum: [FSM, MOCK, COMARCH]
          example:
            fullName: FSM
            address: FSM
            phones: FSM
            email: FSM
            latitude: FSM
            longitude: FSM
            planName: MOCK
            contractedDownloadMbps: MOCK
            contractedUploadMbps: MOCK
    ClientProfileUpdate:
      type: object
      description: Campos editables del módulo de actualización de datos (campo 4).
      properties:
        fullName:
          type: string
        address:
          type: string
        phones:
          type: array
          items:
            type: string
    AccountStatus:
      type: string
      description: >
        Estado de la cuenta en el modelo interno. FSM publica cinco códigos
        (A/S/T/O/P); todo lo demás cae en DESCONOCIDA.
      enum: [ACTIVA, SUSPENDIDA, TERMINADA, ORDENADA, PENDIENTE, DESCONOCIDA]
    AccountStatusCode:
      type: string
      nullable: true
      description: >
        Código crudo de la operadora: A (activo), S (suspendido), T (terminado),
        O (ordenado), P (pendiente). `null` si no vino o no se reconoció.
      enum: [A, S, T, O, P, null]
    ContractStatus:
      type: object
      description: >
        FSM no devuelve "otras cuentas del mismo contrato": `accounts` trae
        exactamente un elemento hasta que exista ese endpoint.
      properties:
        clientName:
          type: string
          description: Nombre del cliente según la orden más reciente. `""` si FSM no lo trajo.
        accounts:
          type: array
          items:
            type: object
            properties:
              accountNumber:
                type: string
              contractId:
                type: string
                nullable: true
                description: >
                  ⚠ Ahora puede ser `null`: FSM no expone contrato, solo órdenes
                  de trabajo.
              status:
                $ref: '#/components/schemas/AccountStatus'
              statusCode:
                $ref: '#/components/schemas/AccountStatusCode'
              statusDescription:
                type: string
                nullable: true
                description: Texto literal de la operadora ("Activo", "Suspendido"…).
              lastWorkOrder:
                type: string
                nullable: true
                description: Orden de trabajo más reciente de `/account/process`.
                example: ORDER/424900/2026
        brand:
          type: string
          enum: [telenews, seteinfo]
          description: Marca efectivamente usada en la consulta.

    # ---- Diagnóstico de red (campos 6, 8-14, 19-21) ------------------
...
    AccountOrders:
      type: object
      properties:
        accountNumber:
          type: string
        brand:
          type: string
          enum: [telenews, seteinfo]
        client:
          type: object
          nullable: true
          description: Datos del cliente según la orden más reciente. `null` si no hay órdenes.
          properties:
            names:
              type: string
              nullable: true
            phoneNumber:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
            address:
              type: string
              nullable: true
            latitude:
              type: number
              format: double
              nullable: true
            longitude:
              type: number
              format: double
              nullable: true
        orders:
          type: array
          description: Ordenadas por `createdAt` descendente. Puede ser `[]`.
          items:
            type: object
            properties:
              workOrder:
                type: string
                example: ORDER/424900/2026
              task:
                type: string
                nullable: true
              state:
                type: string
                nullable: true
              externalProcess:
                type: string
                nullable: true
              cpartyId:
                type: string
                nullable: true
              createdAt:
                type: string
                format: date-time
                nullable: true
              endedAt:
                type: string
                format: date-time
                nullable: true
              finished:
                type: boolean
                description: Equivale a `endedAt != null` (`finish_date is not null`).
              note:
                type: string
                nullable: true
              latitude:
                type: number
                format: double
                nullable: true
              longitude:
                type: number
                format: double
                nullable: true
              address:
                type: string
                nullable: true
...